+++ src/Logger/Writer/AbstractWriter.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Logger\Writer;
 
-use ExtendsSoftware\ExaPHP\Logger\Decorator\DecoratorInterface;
 use ExtendsSoftware\ExaPHP\Logger\Filter\FilterInterface;
 use ExtendsSoftware\ExaPHP\Logger\LogInterface;
 use ExtendsSoftware\ExaPHP\ServiceLocator\Resolver\StaticFactory\StaticFactoryInterface;
@@ -18,13 +17,6 @@ abstract class AbstractWriter implements WriterInterface, StaticFactoryInterface
     private array $filters = [];
 
     /**
-     * Decorators.
-     *
-     * @var DecoratorInterface[]
-     */
-    private array $decorators = [];
-
-    /**
      * Add filter.
      *
      * @param FilterInterface $filter
@@ -36,36 +28,6 @@ abstract class AbstractWriter implements WriterInterface, StaticFactoryInterface
         $this->filters[] = $filter;
 
         return $this;
-    }
-
-    /**
-     * Add decorator.
-     *
-     * @param DecoratorInterface $decorator
-     *
-     * @return AbstractWriter
-     */
-    public function addDecorator(DecoratorInterface $decorator): AbstractWriter
-    {
-        $this->decorators[] = $decorator;
-
-        return $this;
-    }
-
-    /**
-     * Decorate log and return new instance.
-     *
-     * @param LogInterface $log
-     *
-     * @return LogInterface
-     */
-    protected function decorate(LogInterface $log): LogInterface
-    {
-        foreach ($this->decorators as $decorator) {
-            $log = $decorator->decorate($log);
-        }
-
-        return $log;
     }
 
     /**