+++ tests/Logger/LoggerTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace ExtendsSoftware\ExaPHP\Logger;
 
+use ExtendsSoftware\ExaPHP\Logger\Decorator\DecoratorInterface;
 use ExtendsSoftware\ExaPHP\Logger\Priority\PriorityInterface;
 use ExtendsSoftware\ExaPHP\Logger\Writer\WriterInterface;
 use PHPUnit\Framework\TestCase;
@@ -15,14 +16,24 @@ class LoggerTest extends TestCase
      * Test that message will be logged with priority and metadata.
      *
      * @covers \ExtendsSoftware\ExaPHP\Logger\Logger::addWriter()
+     * @covers \ExtendsSoftware\ExaPHP\Logger\Logger::addDecorator()
      * @covers \ExtendsSoftware\ExaPHP\Logger\LoggerWriter::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Logger\LoggerWriter::getWriter()
      * @covers \ExtendsSoftware\ExaPHP\Logger\LoggerWriter::mustInterrupt()
      * @covers \ExtendsSoftware\ExaPHP\Logger\Logger::log()
+     * @covers \ExtendsSoftware\ExaPHP\Logger\Logger::decorate()
      */
     public function testLog(): void
     {
         $priority = $this->createMock(PriorityInterface::class);
+
+        $decorator = $this->createMock(DecoratorInterface::class);
+        $decorator
+            ->expects($this->once())
+            ->method('decorate')
+            ->willReturnCallback(function (LogInterface $log) {
+                return $log->andMetaData('baz', 'qux');
+            });
 
         $writer = $this->createMock(WriterInterface::class);
         $writer
@@ -31,7 +42,7 @@ class LoggerTest extends TestCase
             ->with($this->callback(function (LogInterface $log) use ($priority) {
                 $this->assertSame('Error!', $log->getMessage());
                 $this->assertSame($priority, $log->getPriority());
-                $this->assertSame(['foo' => 'bar'], $log->getMetaData());
+                $this->assertSame(['foo' => 'bar', 'baz' => 'qux'], $log->getMetaData());
 
                 return true;
             }));
@@ -43,6 +54,7 @@ class LoggerTest extends TestCase
         $logger = new Logger();
         $logger
             ->addWriter($writer)
+            ->addDecorator($decorator)
             ->log('Error!', $priority, ['foo' => 'bar']);
     }